register internship refactored, compiles now.
