+++ src/resolver/RegisterInternship.ts
@@ -1,10 +1,5 @@
 import {Db} from "mongodb";
-import {InternshipOffer} from "../schema/InternshipOffer";
-
-export const internshipResolver = async (db: Db) => {
-    const internshipOffers = await db.collection('Internships').find({}).toArray();
-    return internshipOffers.map(internshipOffer => new InternshipOffer(internshipOffer));
-};
+import {InternshipOffer} from "./InternshipOffer";
 
 export const registerInternshipResolver = (db: Db) => async ({input}) => {
     const doc = {
